+++ src/Castle.Facilities.Logging.Log4net.Tests/Log4NetFacilityTests.cs
@@ -12,36 +12,35 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+#if CASTLE_SERVICES_LOGGING
 namespace Castle.Facilities.Logging.Tests
 {
-    using System;
-
-    using Castle.Facilities.Logging.Tests.Classes;
-    using Castle.MicroKernel.Registration;
-    using Castle.Windsor;
+	using System;
+	using System.IO;
+	using Castle.Facilities.Logging.Tests.Classes;
+	using Castle.MicroKernel.Registration;
+	using Castle.Windsor;
+	using log4net;
+	using log4net.Appender;
+	using log4net.Layout;
+	using log4net.Repository.Hierarchy;
 	using NUnit.Framework;
 
-	/// <summary>
-	/// Summary description for ConsoleFacitlyTest.
-	/// </summary>
 	[TestFixture]
-	public class NullFacilityTest : BaseTest
+	public class Log4NetFacilityTests : Log4NetBaseTest
 	{
 		private IWindsorContainer container;
 
 		[SetUp]
 		public void Setup()
 		{
-			container = base.CreateConfiguredContainer(LoggerImplementation.Null);
+			container = base.CreateConfiguredContainerLog4Net(Log4netFacility.LoggerImplementation.ExtendedLog4Net);
 		}
 
 		[TearDown]
 		public void Teardown()
 		{
-			if (container != null)
-			{
-				container.Dispose();
-			}
+			container.Dispose();
 		}
 
 		[Test]
@@ -51,6 +50,15 @@ namespace Castle.Facilities.Logging.Tests
 			SimpleLoggingComponent test = container.Resolve<SimpleLoggingComponent>("component");
 
 			test.DoSomething();
+
+			String expectedLogOutput = String.Format("[INFO ] [{0}] - Hello world" + Environment.NewLine, typeof(SimpleLoggingComponent).FullName);
+			MemoryAppender memoryAppender = ((Hierarchy) LogManager.GetRepository()).Root.GetAppender("memory") as MemoryAppender;
+			TextWriter actualLogOutput = new StringWriter();
+			PatternLayout patternLayout = new PatternLayout("[%-5level] [%logger] - %message%newline");
+			patternLayout.Format(actualLogOutput, memoryAppender.GetEvents()[0]);
+
+			Assert.AreEqual(expectedLogOutput, actualLogOutput.ToString());
 		}
 	}
 }
+#endif