+++ src/Castle.Facilities.WcfIntegration.Tests/Service/ServiceHostFixture.cs
@@ -21,7 +21,7 @@ namespace Castle.Facilities.WcfIntegration.Tests
 	using System.ServiceModel.Description;
 	using System.ServiceModel.Discovery;
 	using Castle.Core.Resource;
-	using Castle.Facilities.Logging;
+	using Castle.Facilities.Logging.Log4netFacility;
 	using Castle.Facilities.WcfIntegration.Behaviors;
 	using Castle.Facilities.WcfIntegration.Demo;
 	using Castle.Facilities.WcfIntegration.Tests.Behaviors;
@@ -1019,7 +1019,7 @@ namespace Castle.Facilities.WcfIntegration.Tests
 
 		protected IWindsorContainer RegisterLoggingFacility(IWindsorContainer container)
 		{
-			var logging = new LoggingFacility(LoggerImplementation.ExtendedLog4net);
+			var logging = new LoggingFacility(LoggerImplementation.ExtendedLog4Net);
 			container.AddFacility(logging);
 
 			memoryAppender = new MemoryAppender();